Add files via upload
The index.js file includes a switch statement to determine how to handle a randomly generated number between 1 and 3 for the computer's choice in the game. The tutorial used a series of if statements, but I commented those out to implement a switch statement. A little more easier since we know we are dealing with only 3 possible outcomes